client/requirements.txt
streamlit==1.48.0
python-dotenv==1.1.1
nest-asyncio==1.6.0
httpx>=0.28.1

# Core AI/ML packages
langchain==0.3.20
langchain-openai>=0.0.3
openai>=1.99.3

# OpenAI Agents SDK
openai-agents>=0.2.5
mcp>=1.12.4

# Utilities
pandas==2.3.1